moving Vocab Factory functions out of experimental
